* Rime communication schemes work, examples found in example/rime * Examples cleaned and described with READMEs * User application may embed a contiki-conf-app.h file allowing custom activation of modules such as uip or rime. See the different examples for details. * Added a script for launching tunslip easily.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/senstools/trunk@163 e59e8366-ff4a-0410-815a-bc13a6274fd4

+++ software/OS/Contiki/cpu/msp430/mtarch.c
@@ -65,7 +65,7 @@ mtarch_start(struct mtarch_thread *t,
   /* Space for registers. */
   t->sp -= 11;
   
-  *t->sp = data;
+  *t->sp = (unsigned short) data;
 }
 /*--------------------------------------------------------------------------*/
 static unsigned short *sptmp;
